+++ src/main/java/ru/practicum/shareit/requests/ItemRequestService.java
@@ -1,5 +1,6 @@
 package ru.practicum.shareit.requests;
 
+import org.springframework.data.domain.Page;
 import ru.practicum.shareit.item.ItemDtoForRequest;
 
 import java.util.List;
@@ -7,6 +8,6 @@ import java.util.List;
 public interface ItemRequestService {
     ItemRequestDto safe(Long userId, ItemRequestDto itemRequestDto);
     List<ItemRequestDto> findRequestByUserId(Long userId);
-    List<ItemRequestDto> findRequestByParam(Long userId, Long from, Long size);
+    List<ItemRequestDto> findRequestByParam(Long userId, int from, int size);
     ItemRequestDto findById(Long userId, Long requestId);
 }
